Changed pins variable type from int8_t to int16_t.
So that we can accommodate pin number greater than 128. Fixes #60.

+++ src/Adafruit_MCPSRAM.cpp
@@ -13,8 +13,8 @@
                 @param  cs chip select pin
 */
 /**************************************************************************/
-Adafruit_MCPSRAM::Adafruit_MCPSRAM(int8_t mosi, int8_t miso, int8_t sck,
-                                   int8_t cs) {
+Adafruit_MCPSRAM::Adafruit_MCPSRAM(int16_t mosi, int16_t miso, int16_t sck,
+                                   int16_t cs) {
   _mosi = mosi;
   _miso = miso;
   _sck = sck;
@@ -29,7 +29,7 @@ Adafruit_MCPSRAM::Adafruit_MCPSRAM(int8_t mosi, int8_t miso, int8_t sck,
                 @param spi the SPI bus to use
 */
 /**************************************************************************/
-Adafruit_MCPSRAM::Adafruit_MCPSRAM(int8_t cs, SPIClass *spi) {
+Adafruit_MCPSRAM::Adafruit_MCPSRAM(int16_t cs, SPIClass *spi) {
   _cs = cs;
   _spi = spi;
   hwSPI = true;
