+++ src/panels/ThinkInk_213_Mono_M21.h
@@ -6,12 +6,12 @@
 
 class ThinkInk_213_Mono_M21 : public Adafruit_UC8151D {
 public:
-  ThinkInk_213_Mono_M21(int8_t SID, int8_t SCLK, int8_t DC, int8_t RST,
-                        int8_t CS, int8_t SRCS, int8_t MISO, int8_t BUSY = -1)
+  ThinkInk_213_Mono_M21(int16_t SID, int16_t SCLK, int16_t DC, int16_t RST,
+                        int16_t CS, int16_t SRCS, int16_t MISO, int16_t BUSY = -1)
       : Adafruit_UC8151D(212, 104, SID, SCLK, DC, RST, CS, SRCS, MISO, BUSY){};
 
-  ThinkInk_213_Mono_M21(int8_t DC, int8_t RST, int8_t CS, int8_t SRCS,
-                        int8_t BUSY = -1, SPIClass *spi = &SPI)
+  ThinkInk_213_Mono_M21(int16_t DC, int16_t RST, int16_t CS, int16_t SRCS,
+                        int16_t BUSY = -1, SPIClass *spi = &SPI)
       : Adafruit_UC8151D(212, 104, DC, RST, CS, SRCS, BUSY, spi){};
 
   void begin(thinkinkmode_t mode = THINKINK_MONO) {
